debug session with amir hosein code cleaning

+++ InternetBank.Core/Domain/Entities/Account.cs
@@ -4,8 +4,7 @@ namespace InternetBank.Core.Domain.Entities;
 
 public class Account
 {
-    //sets as a random number
-    public int AccountId { get; set; }
+    public long AccountId { get; set; }
     //sets as a random number
     public string AccountNumber { get; set; }
     //sets as a random number
@@ -15,7 +14,6 @@ public class Account
     //sets as a random number
     public DateTime ExpireDate { get; set; }
     public AccountType accountType { get; set; }
-    
     //sets as a random number
     public string AccountStaticPassword { get; set; }
     public string Amount { get; set; }
